split EntityPriorityQueue stuff into separate file
