+++ assignment-client/src/entities/EntityTreeSendThread.h
@@ -12,80 +12,15 @@
 #ifndef hifi_EntityTreeSendThread_h
 #define hifi_EntityTreeSendThread_h
 
-#include <queue>
-
 #include "../octree/OctreeSendThread.h"
 
-#include <AACube.h>
-
-#include "EntityTreeElement.h"
-
-const float SQRT_TWO_OVER_TWO = 0.7071067811865;
-const float DEFAULT_VIEW_RADIUS = 10.0f;
+#include "EntityPriorityQueue.h"
 
 class EntityNodeData;
 class EntityItem;
 
-
-class ConicalView {
-public:
-    ConicalView() {}
-    ConicalView(const ViewFrustum& viewFrustum) { set(viewFrustum); }
-    void set(const ViewFrustum& viewFrustum);
-    float computePriority(const AACube& cube) const;
-    float computePriority(const EntityItemPointer& entity) const;
-private:
-    glm::vec3 _position { 0.0f, 0.0f, 0.0f };
-    glm::vec3 _direction { 0.0f, 0.0f, 1.0f };
-    float _sinAngle { SQRT_TWO_OVER_TWO };
-    float _cosAngle { SQRT_TWO_OVER_TWO };
-    float _radius { DEFAULT_VIEW_RADIUS };
-};
-
-class PrioritizedEntity {
-public:
-    PrioritizedEntity(EntityItemPointer entity, float priority) : _weakEntity(entity), _priority(priority) { }
-    float updatePriority(const ConicalView& view);
-    EntityItemPointer getEntity() const { return _weakEntity.lock(); }
-    float getPriority() const { return _priority; }
-
-    class Compare {
-    public:
-        bool operator() (const PrioritizedEntity& A, const PrioritizedEntity& B) { return A._priority < B._priority; }
-    };
-    friend class Compare;
-
-private:
-    EntityItemWeakPointer _weakEntity;
-    float _priority;
-};
-
-class VisibleElement {
-public:
-    EntityTreeElementPointer element;
-    ViewFrustum::intersection intersection { ViewFrustum::OUTSIDE };
-};
-
-class Fork {
-public:
-    Fork(EntityTreeElementPointer& element);
-
-    void getNextVisibleElementFirstTime(VisibleElement& next, const ViewFrustum& view);
-    void getNextVisibleElementAgain(VisibleElement& next, const ViewFrustum& view, uint64_t lastTime);
-    void getNextVisibleElementDifferential(VisibleElement& next, const ViewFrustum& view, const ViewFrustum& lastView, uint64_t lastTime);
-
-    int8_t getNextIndex() const { return _nextIndex; }
-    void initRootNextIndex() { _nextIndex = -1; }
-
-protected:
-    EntityTreeElementWeakPointer _weakElement;
-    int8_t _nextIndex;
-};
-
-using EntityPriorityQueue = std::priority_queue< PrioritizedEntity, std::vector<PrioritizedEntity>, PrioritizedEntity::Compare >;
-
-
 class EntityTreeSendThread : public OctreeSendThread {
+
 public:
     EntityTreeSendThread(OctreeServer* myServer, const SharedNodePointer& node);
 
@@ -107,7 +42,7 @@ private:
     ViewFrustum _currentView;
     ViewFrustum _completedView;
     ConicalView _conicalView; // optimized view for fast priority calculations
-    std::vector<Fork> _traversalPath;
+    std::vector<TraversalWaypoint> _traversalPath;
     std::function<void (VisibleElement&)> _getNextVisibleElementCallback { nullptr };
     std::function<void (VisibleElement&)> _scanNextElementCallback { nullptr };
     uint64_t _startOfCompletedTraversal { 0 };
